Embeddings to pending request
