+++ tests/Fixtures/ollama/structured-1.json
@@ -1,1 +1,1 @@
-{"id":"chatcmpl-480","object":"chat.completion","created":1737064422,"model":"qwen2.5:14b","system_fingerprint":"fp_ollama","choices":[{"index":0,"message":{"role":"assistant","content":"{\n    \"weather\": \"moderate temperature with 70ºF\",\n    \"game_time\": \"3pm\",\n    \"coat_required\": false\n}"},"finish_reason":"stop"}],"usage":{"prompt_tokens":187,"completion_tokens":34,"total_tokens":221}}
+{"id":"chatcmpl-606","object":"chat.completion","created":1737066390,"model":"qwen2.5:14b","system_fingerprint":"fp_ollama","choices":[{"index":0,"message":{"role":"assistant","content":"{\n    \"weather\": \"moderate temperature\",\n    \"game_time\": \"3pm\",\n    \"coat_required\": false\n}"},"finish_reason":"stop"}],"usage":{"prompt_tokens":187,"completion_tokens":28,"total_tokens":215}}
