+++ tests/Providers/Ollama/StructuredTest.php
@@ -12,7 +12,7 @@ use EchoLabs\Prism\Schema\StringSchema;
 use Tests\Fixtures\FixtureResponse;
 
 it('returns structured output', function (): void {
-    FixtureResponse::recordResponses('v1/chat/completions', 'ollama/structured');
+    FixtureResponse::fakeResponseSequence('v1/chat/completions', 'ollama/structured');
 
     $schema = new ObjectSchema(
         'output',
